Move and add code excerpts for new loops page
