+++ examples/language/dart_test.yaml
@@ -1,0 +1,1 @@
+include: ../dart_test_base_browser.yaml
