+++ examples/language/pubspec.yaml
@@ -1,0 +1,12 @@
+name: examples
+description: dart.dev example code.
+
+environment:
+  sdk: ^3.0.0-0
+
+dependencies:
+  examples_util: { path: ../util }
+
+dev_dependencies:
+  lints: ^2.1.0
+  test: ^1.24.2
